requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
loguru==0.7.2
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.7
python-dotenv==1.0.0
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
jinja2==3.1.2
aiofiles==23.2.1

